Extract aggregation sinks and results-cache connect helper into their own files.
Move AggregationSink/StdoutSink/ResultsCacheSink out of OutputHandlerImpl
into AggregationSink.{hpp,cpp} (they are a distinct abstraction from
search::OutputHandler), and lift the shared connect_to_results_cache helper
into ResultsCacheUtils.hpp so both ResultsCacheOutputHandler and ResultsCacheSink
share one definition instead of a file-private one. AggregationOutputHandler
stays in OutputHandlerImpl since it is a search::OutputHandler.
